Trait pallet_linkable_tree_rpc::LinkableTreeRpcApiServer

pub trait LinkableTreeRpcApiServer<BlockHash, E, C, L>: Sized + Send + Sync + 'static {
    fn get_neighbor_roots(
        &self,
        tree_id: u32,
        at: Option<BlockHash>
    ) -> RpcResult<Vec<E>>;
    fn get_neighbor_edges(
        &self,
        tree_id: u32,
        at: Option<BlockHash>
    ) -> RpcResult<Vec<EdgeMetadata<C, E, L>>>;

    fn into_rpc(self) -> RpcModule<Self>
    where
        BlockHash: Send + Sync + 'static + DeserializeOwned,
        E: Send + Sync + 'static + Serialize,
        C: Send + Sync + 'static + Serialize,
        L: Send + Sync + 'static + Serialize,
    { ... }
}

Server trait implementation for the LinkableTreeRpcApi RPC API.

Required Methods

fn get_neighbor_roots(
&self,
tree_id: u32,
at: Option<BlockHash>
) -> RpcResult<Vec<E>>

Get the Linkable Tree neighbor roots.

Returns the (full) Vec of the neighbor roots

fn get_neighbor_edges(
&self,
tree_id: u32,
at: Option<BlockHash>
) -> RpcResult<Vec<EdgeMetadata<C, E, L>>>

Get the Linkable Tree neighbor edges.

Returns the (full) Vec of the neighbor edge metadata

Provided Methods

fn into_rpc(self) -> RpcModule<Self>where
BlockHash: Send + Sync + 'static + DeserializeOwned,
E: Send + Sync + 'static + Serialize,
C: Send + Sync + 'static + Serialize,
L: Send + Sync + 'static + Serialize,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.